requirement.txt
discord.py[voice]>=2.3.0
python-dotenv==1.0.0
yt-dlp>=2024.1.1
async-timeout==4.0.2
pillow>=9.3.0
aiohttp>=3.9.0
PyNaCl>=1.4.0
yandex-music>=2.0.0
